refactor: move source files into src/lemongrass/ package
