+++ src/lemongrass/__init__.py
@@ -1,0 +1,1 @@
+"""Lemongrass motorsport lap timing tool."""
