+++ tests/test_pisugar_monitor.py
@@ -1,16 +1,9 @@
 # pylint: disable=missing-module-docstring,missing-class-docstring,missing-function-docstring
 import base64
-import importlib.util
 import json
-import pathlib
 from unittest.mock import mock_open, patch
 
-_spec = importlib.util.spec_from_file_location(
-    "pisugar_monitor",
-    pathlib.Path(__file__).parent.parent / "pisugar-monitor.py",
-)
-_mod = importlib.util.module_from_spec(_spec)
-_spec.loader.exec_module(_mod)
+import lemongrass.pisugar_monitor as _mod
 
 token_expiry = _mod.token_expiry
 read_credentials = _mod.read_credentials
